fix(success): rename folder
